// services/api/src/learning-orchestration/enums/learning-flow-state.enum.ts
export enum LearningFlowState {

  TCC_PENDING = 'TCC_PENDING',

  TCC_READY = 'TCC_READY',

  DSC_PENDING = 'DSC_PENDING',

  DSC_READY = 'DSC_READY',

  CONTENT_PENDING = 'CONTENT_PENDING',

  CONTENT_READY = 'CONTENT_READY',

  LESSON_READY = 'LESSON_READY',

  LESSON_IN_PROGRESS = 'LESSON_IN_PROGRESS',

  LESSON_COMPLETED = 'LESSON_COMPLETED',

  CBT_READY = 'CBT_READY',

  CBT_IN_PROGRESS = 'CBT_IN_PROGRESS',

  CBT_COMPLETED = 'CBT_COMPLETED',

  STARS_ELIGIBLE = 'STARS_ELIGIBLE',

  QUESTION_MODE = 'QUESTION_MODE',

  COMPLETED = 'COMPLETED',

}